% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVecLMs.R
\name{calcVecLMs}
\alias{calcVecLMs}
\title{Create a linear regression matrix.}
\usage{
calcVecLMs(bin_data, use_slurm = F, job_finished = F, slurmjob = NULL,
  n_nodes = NULL, cpus_on_each_node = 1, memory_per_node = "2g",
  walltime = "4:00:00", n_cores = 1)
}
\arguments{
\item{bin_data}{The parent matrix, with columns to have linear regression performed on them.}

\item{use_slurm}{Paralleize over a number of slurm HPC jobs? If false, the program will simply run locally.}

\item{slurmjob}{the slurm job object produced by rslurm::slurm_apply(), after running the function initially.}
}
\value{
The output matrix, or if using slurm, the slurm job object (which should be saved as an rds file and reloaded when creating the output matrix).
}
\description{
Creates a matrix of linear regression p-values, log transformed from every combination of columns in the parent matrix.
}
\examples{
\dontrun{
#small example
#bin_data<-matrix(runif(5*5),ncol=5)
#full_matrix<-calcVecLMs(bin_data)
}
}
\keyword{linear}
\keyword{lm}
\keyword{matrix}
\keyword{regression}
