% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChromosomalMatrixSet.R
\name{createChromosomalMatrixSet}
\alias{createChromosomalMatrixSet}
\title{Create chromosomal interaction matrices for HiCNV shiny application.}
\usage{
createChromosomalMatrixSet(whole_genome_mat, output_dir, prefix = "nbl_")
}
\arguments{
\item{whole_genome_mat}{The matrix containing all of the data, from which the individual matrices will be split.}

\item{output_dir}{the folder where the matrices in RData format, will be written.}

\item{prefix}{filename prefix for individual matrices. Default: "nbl_"}
}
\value{
The list of files already written to disk, with full filenames and paths.
}
\description{
Takes a linear regression matrix and sets infinites to a finite value, and changes the sign to match the sign of the correlation for each value.
}
\keyword{Interaction}
\keyword{matrix}
