% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divisors.R
\name{divisors}
\alias{divisors}
\title{List of Divisors}
\usage{
divisors(n)
}
\arguments{
\item{n}{an integer whose divisors will be generated.}
}
\value{
Returns a vector integers.
}
\description{
Generates a list of divisors of an integer number.
Identical to the same function within the numbers package.
The code has been modified from the numbers package,
 following GPL 3.0 guidelines on 3/30/2022, section 5.
 Reference for GPL v3.0 LICENSE:
 https://www.gnu.org/licenses/gpl-3.0.en.html.
}
\examples{
divisors(1)          # 1
divisors(2)          # 1 2
divisors(3)          # 1 2 3
divisors(2^5)        # 1  2  4  8 16 32
divisors(1000)       # 1  2  4  5  8 10 ... 100 125 200 250 500 1000
divisors(1001)       # 1  7 11 13 77 91 143 1001
}
\seealso{
[numbers::divisors()]
}
\keyword{divisors}
\keyword{numbers}
