% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnnotationMatrix.R
\name{getAnnotationMatrix}
\alias{getAnnotationMatrix}
\title{Get the genes in the genomic ranges indicated by the row and column labels.}
\usage{
getAnnotationMatrix(genomic_matrix, prot_only = T, sequential = F)
}
\arguments{
\item{genomic_matrix}{A matrix with row and column names of the format chr1_100_200 (chr,start,end)}

\item{prot_only}{Inlcude only the protein coding genes from ensembl?}

\item{sequential}{Turn off parallelism with DoMC?}
}
\value{
concatenated_gene_matrix A matrix with row and column genes
}
\description{
Gets the genes in the ranges within each cell of the matrix.
}
\keyword{genomic}
\keyword{matrix}
