% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head.mat.R
\name{head.mat}
\alias{head.mat}
\title{Gets a small piece of a matrix (top left corner) for viewing, rather than pulling the first n rows.}
\usage{
head.mat(mat, n = 6L)
}
\arguments{
\item{mat}{A matrix.}

\item{n}{The length and width of the piece to view.}
}
\value{
averaged_matrix a small matrix of size n.
}
\description{
Gives a small square of a matrix to get an idea of content rather than grabbing the entire row.
When this row is thousands of numbers long, this can be a problem.
}
\keyword{average}
\keyword{downsample}
\keyword{edges}
\keyword{matrix}
\keyword{rescale}
