% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiCNVserver.R
\name{HiCNVserver}
\alias{HiCNVserver}
\title{Server component of the HiCNV plotly shiny application.}
\arguments{
\item{session}{The shiny session object for the application.}

\item{input}{shiny server input}

\item{output}{shiny server output}
}
\value{
None
}
\description{
Server function of the HiCNV shiny application. run with runHiCNVShiny
}
\examples{
\dontrun{
runHiCNVShiny()
}
}
\keyword{CNV}
\keyword{heatmap}
\keyword{plotly}
\keyword{shiny}
