% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mathead.R
\name{mathead}
\alias{mathead}
\title{Gets a small piece of a matrix (top left corner) for viewing, rather than pulling the first n rows.}
\usage{
mathead(mat, n = 6L)
}
\arguments{
\item{mat}{A matrix.}

\item{n}{The length and width of the piece to view.}
}
\value{
averaged_matrix a small matrix of size n.
}
\description{
Gives a small square of a matrix to get an idea of content rather than grabbing the entire row.
When this row is thousands of numbers long, this can be a problem.
}
\examples{
load(system.file("extdata","nbl_result_matrix_sign_small.rda",package = "CNVScope"))
mathead(nbl_result_matrix_sign_small)
}
\keyword{average}
\keyword{downsample}
\keyword{edges}
\keyword{matrix}
\keyword{rescale}
