% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbl_result_matrix_sign_small.R
\name{nbl_result_matrix_sign_small}
\alias{nbl_result_matrix_sign_small}
\title{Neuroblastoma sample CNV relationship matrix}
\format{
A matrix with 25 rows and 25 variables
}
\source{
\url{https://gdc.cancer.gov/}
}
\description{
The first 25 Mb of chromosome 1,
neuroblastoma copy number signed relation matrix.
}
