% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCNVScopeLocal.R
\name{runCNVScopeLocal}
\alias{runCNVScopeLocal}
\title{Runs the CNVScope plotly shiny application.}
\usage{
runCNVScopeLocal()
}
\value{
none. Runs the application if the correct files are present.
}
\description{
Runs the interactive suite of tools locally.
}
\examples{
\dontrun{
CNVScope::runCNVScopeLocal()
}
}
\keyword{CNV}
\keyword{heatmap}
\keyword{plotly}
\keyword{shiny}
