% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHiCNVShiny.R
\name{runHiCNVShiny}
\alias{runHiCNVShiny}
\title{Runs the HiCNV plotly shiny application.}
\arguments{
\item{baseurl}{the url of the source files for the application (e.g. the contents of plotly_dashboard_ext). This will be pulled from remotely.}

\item{basefn}{the linux file path of the same source files.}
}
\value{
none. Runs the application if the correct files are present.
}
\description{
Runs the interactive suite of tools locally or on a server if called in a script file (e.g. App.R).
}
\examples{
\dontrun{
runHiCNVShiny()
}
}
\keyword{CNV}
\keyword{heatmap}
\keyword{plotly}
\keyword{shiny}
