% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signedRescale.R
\name{signedRescale}
\alias{signedRescale}
\title{Rescale positive and negative data, preserving sign information.}
\usage{
signedRescale(matrix)
}
\arguments{
\item{matrix}{A matrix to be transformed}
}
\value{
transformedmatrix A transformed matrix.
}
\description{
Performs a signed rescale on the data, shrinking the negative and positive ranges into the [0,1] space, such that negative is always less than 0.5 and positive is always greater.
}
\keyword{matrix}
\keyword{negative}
\keyword{positive}
\keyword{rescale}
\keyword{signed}
