% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCNVKernelProbDist.R
\name{calcCNVKernelProbDist}
\alias{calcCNVKernelProbDist}
\title{Calculate the probability distribution of CNV concordance events with a fast kernel}
\usage{
calcCNVKernelProbDist(submatrix = NULL, win = 5)
}
\arguments{
\item{submatrix}{A matrix of CNV data in an intrachromosomal region (e.g. chr1 vs chr1 or chr5 vs chr5)}

\item{win}{a window size for the matrix that calculates the windowed average using the kernel function}
}
\description{
This function produces several matrices, including a Z-score matrix from a matrix of the same size and a percentile matrix of these Z-scores
}
\examples{
set.seed(303)
mat<-matrix(data=runif(n = 25),nrow=5,ncol=5)
calcCNVKernelProbDist(matrix)
mat_prob_dist<-calcCNVKernelProbDist(mat)
mat_prob_dist
ComplexHeatmap::Heatmap(mat_prob_dist$zscore_matrix,cluster_columns = F,cluster_rows = F,show_column_names = F,show_row_names = F,column_title = "z scores") + 
ComplexHeatmap::Heatmap(mat_prob_dist$percentile_matrix,cluster_columns = F,cluster_rows = F,show_column_names = F,show_row_names = F,column_title = "percentile scores") + 
ComplexHeatmap::Heatmap(mat_prob_dist$original_matrix,cluster_columns = F,cluster_rows = F,show_column_names = F,show_row_names = F,column_title = "original data")
}
\keyword{CNV}
\keyword{concordance}
\keyword{distribution}
\keyword{fast}
\keyword{kernel}
\keyword{probability}
