% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createChromosomalMatrixSet.R
\name{createChromosomalMatrixSet}
\alias{createChromosomalMatrixSet}
\title{Create chromosomal interaction matrices for CNVScope shiny application.}
\usage{
createChromosomalMatrixSet(whole_genome_mat, output_dir = NULL,
  prefix = "nbl_")
}
\arguments{
\item{whole_genome_mat}{The matrix containing all of the data, from which the individual matrices will be split.}

\item{output_dir}{the folder where the matrices in RData format, will be written.}

\item{prefix}{filename prefix for individual matrices. Default: "nbl_"}
}
\value{
The list of files already written to disk, with full filenames and paths.
}
\description{
Takes a linear regression matrix and sets infinites to a finite value, and changes the sign to match the sign of the correlation for each value.
}
\examples{
#examples for this function would be too large to 
#include and should be run on an HPC machine node.
#illustration of this process is shown clearly in 
#the vignette and can be done if a user properly
#follows the instructions.
# The function is intended to be run on a whole interactome matrix (chr1-X).
}
\keyword{Interaction}
\keyword{matrix}
