% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_genomic_matrix.R
\name{downsample_genomic_matrix}
\alias{downsample_genomic_matrix}
\title{Rescale positive and negative data, preserving sign information.}
\arguments{
\item{whole_matrix}{A matrix to be downsampled, on a single chromosome}

\item{downsamplefactor}{A factor by which to reduce the matrix. Must be something that both the row and columns can be divisible by.}

\item{singlechromosome}{Single chromosome mode; Multi-chromosome not yet implemented (leave T)}
}
\value{
whole_matrix_dsamp A downsampled matrix.
}
\description{
Downsamples a matrix by a specified factor.
}
\examples{
test_mat<-matrix(runif(81),nrow=9)
colnames(test_mat)<-paste0("chr1_",1:ncol(test_mat))
rownames(test_mat)<-paste0("chr1_",1:nrow(test_mat))
downsample_genomic_matrix(whole_matrix=test_mat,downsamplefactor=3,singlechromosome=T)
}
\keyword{matrix}
\keyword{negative}
\keyword{positive}
\keyword{rescale}
\keyword{signed}
