% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formSampleMatrixFromRawGDCData.R
\name{formSampleMatrixFromRawGDCData}
\alias{formSampleMatrixFromRawGDCData}
\title{Form sample matrix from GDC copy number data files.}
\arguments{
\item{tcga_files}{GDC files to be read}

\item{format}{file format, TCGA or TARGET.}

\item{binsize}{the binsize, in base pairs (default 1Mb or 1e6).  This value provides a good balance of resolution and speed with memory sensitive applications.}

\item{freadskip}{the number of lines to skip in the GDC files, typically 14 (the first 13 lines are metadata and the first is a blank line in NBL data). Adjust as needed.}

\item{debug}{debug mode enable (allows specific breakpoints to be checked).}

\item{chromosomes}{A vector of chromosomes to be used. Defaults to chr1-chrX,
but others can be added e.g. chrY or chrM for Y chromosome or mitochondrial DNA.
 Format expected is a character vector, e.g. c("chr1", "chr2", "chr3").}

\item{sample_pat}{Pattern used to extract sample name from filename.
Use "" to use the filename.}

\item{sample_col}{The name of the sample column (for custom format input).}

\item{chrlabel}{The name of the chromosome column (for custom format input).}

\item{startlabel}{The name of the start column (for custom format input).}

\item{endlabel}{The name of the end column (for custom format input).}
}
\value{
A dataframe containing the aggregated copy number values,

 based on the parameters provided.
}
\description{
Reads a GDC segmetnation files, adds sample information, and forms a data matrix of samples and bins of a specified size.
}
\examples{
#Pipeline examples would be too large to include in package checks.
#please see browseVignettes("CNVScope") for a demonstration.

}
\keyword{GDC}
\keyword{segmentation}
