% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formSampleMatrixFromRawGDCData.R
\name{formSampleMatrixFromRawGDCData}
\alias{formSampleMatrixFromRawGDCData}
\title{Form sample matrix from GDC low-pass segmentation datafiles.}
\arguments{
\item{tcga_files}{GDC files to be read}

\item{format}{file format, TCGA or TARGET.}

\item{binsize}{the binsize, in base pairs (default 1Mb or 1e6).  This value provides a good balance of resolution and speed with memory sensitive applications.}

\item{freadskip}{the number of lines to skip in the GDC files, typically 14 (the first 13 lines are metadata and the first is a blank line in NBL data). Adjust as needed.}
}
\value{
sample_aggregated_segvals A dataframe containing the aggregated segmentation values, based on the parameters provided.
}
\description{
Reads a GDC segmetnation files, adds sample information, and forms a data matrix of samples and bins of a specified size.
}
\examples{
#Pipeline examples would be too large to include in package checks.
#please see browseVignettes("HiCNV") for a demonstration.
}
\keyword{GDC}
\keyword{segmentation}
