% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInterchromosomalInteractivePlot.R
\name{getInterchromosomalInteractivePlot}
\alias{getInterchromosomalInteractivePlot}
\title{Create an HTML widget for use in shiny or webshot for a given pair of chromosomes.}
\usage{
getInterchromosomalInteractivePlot(whole_matrix, chrom1, chrom2)
}
\arguments{
\item{whole_matrix}{the large, whole genomic matrix from which the submatrix is taken}
}
\value{
An HTML widget.
}
\description{
This function requires a matrix with genomic coordinates in the row and column names, and produces a heatmap with a tooltip
}
\examples{
all_conc<-data.table::fread(paste0(groupdir,"dalgleishjl/hicnv/all_concondace_neg_log10_pvalue_seg.txt"),data.table = F)
rownames(all_conc)<-all_conc$pos
all_conc_cleaned<-all_conc[,5:ncol(all_conc)]
common_coords<-intersect(rownames(all_conc_cleaned),colnames(all_conc_cleaned))
all_conc_cleaned_common_coords<-all_conc_cleaned[common_coords,common_coords]
chromosomes<-paste0("chr",c(seq(1:22),"X"),"_")
chrom.pairs<-expand.grid(1:length(chromosomes),1:length(chromosomes))
getInterchromosomalInteractivePlot()
}
\keyword{CNV}
\keyword{HTML}
\keyword{data.table}
\keyword{heatmap}
\keyword{readr}
\keyword{widget}
