% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInterchromosomalInteractivePlot.R
\name{getInterchromosomalInteractivePlot}
\alias{getInterchromosomalInteractivePlot}
\title{Create an HTML widget for use in shiny or webshot for a given pair of chromosomes.}
\arguments{
\item{whole_matrix}{the large, whole genomic matrix from which the submatrix is taken (rows)}

\item{chrom1}{The first chromsome used for the map (columns).}

\item{chrom2}{The second chromsome used for a map axis.}
}
\value{
An HTML widget.
}
\description{
This function requires a matrix with genomic coordinates in the row and column names, and produces a heatmap with a tooltip
}
\examples{
\dontrun{
load(system.file("extdata","nbl_result_matrix_sign_small.rda",package = "HiCNV")) 
getInterchromosomalInteractivePlot(whole_matrix=nbl_result_matrix_sign_small,chrom1=1,
chrom2=1)
}
}
\keyword{CNV}
\keyword{HTML}
\keyword{data.table}
\keyword{heatmap}
\keyword{readr}
\keyword{widget}
