% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVScopeserver.R
\name{CNVScopeserver}
\alias{CNVScopeserver}
\title{Server component of the CNVScope plotly shiny application.}
\arguments{
\item{session}{The shiny session object for the application.}

\item{input}{shiny server input}

\item{output}{shiny server output}
}
\value{
None
}
\description{
Server function of the CNVScope shiny application. run with runCNVScopeShiny
}
\examples{
\dontrun{
runCNVScopeShiny()
}
}
\keyword{CNV}
\keyword{heatmap}
\keyword{plotly}
\keyword{shiny}
