% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBreakpointBed.R
\name{importBreakpointBed}
\alias{importBreakpointBed}
\title{Import a breakpoint BED file.}
\arguments{
\item{breakpoint_fn}{the filename of the breakpoint bed file}
}
\value{
a Genomic Interactions Object
}
\description{
Imports a BED file with breakpoints or other interactions, in a dual position format.
}
\examples{
importBreakpointBed(breakpoint_fn = system.file("extdata","sample_breakpoints.bed",package = "HiCNV"))
}
\keyword{bed}
