% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCNVScopeShiny.R
\name{runCNVScopeShiny}
\alias{runCNVScopeShiny}
\title{Runs the CNVScope plotly shiny application.}
\usage{
runCNVScopeShiny(baseurl = NULL, basefn = NULL, debug = F)
}
\arguments{
\item{baseurl}{the url of the source files for the application (e.g. the contents of plotly_dashboard_ext). This will be pulled from remotely.}

\item{basefn}{the linux file path of the same source files.}

\item{debug}{Enable debugging output.}
}
\value{
none. Runs the application if the correct files are present.
}
\description{
Runs the interactive suite of tools locally or on a server if called in a script file (e.g. App.R).
}
\examples{
\dontrun{
runCNVScopeShiny()
}
}
\keyword{CNV}
\keyword{heatmap}
\keyword{plotly}
\keyword{shiny}
